requirements.txt
apache-libcloud
pyyaml
requests
